Update dependencies after upstream changes
The haskell-grpc-native libs have updated their dependency bounds. A new version of `http2-client` has
been published, but unfortunately the lower bounds on `tls` is currently overly-restrictive and
cannot be built with `http-client-grpc`. Until a meta-revision can be published, we depend on a fork.

The update to `proto-lens-protobuf-types` is not necessary, just keeping things up to date.